requirements.txt
python-telegram-bot==21.6
httpx==0.27.0
anthropic==0.42.0